+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added command /shmarkplayer (marking a player with yellow color)
 - Added highlight slayer miniboss in blue color
 - Added option to hide the death messages of other players, except for players who are close to the player, inside dungeon or doing a Kuudra fight.
+- Added highlight deathmites in dungeon in red color
 
 ### Fixes
 - Fixed small bazaar Message Hider